Reorganize the game request layout slightly
The initial iteration of this had us putting all of the game requests
inline in the top level, but that is a little overwhelming to look at
in that file.

So, this utilizes the same sub-structure that is used for the overall
application into the games section to allow for things to be a little
nicer looking.