Update the game lookup code to be more juicy
The previous iteration of the game lookup DB function would, given a
list of string or number values, try to look up those games and return
back a list of objects with just the id and slug, and nothing else.

For our purposes in the UI we need to be more functional, and allow us
to get the name of the game as well.

This new version of the lookup does that, and can also provide an image
for the game as well.

In addition the query code was tightened up to not waste writes and
reads in D1 trying to query both slugs and id values every time; now it
will only make those queries when it needs to.