Fix session report listing to use the new lookup
The session report list functionality uses the internal game lookup in
the DB to figure what games to look for session on.

When we adjusted that method to return single objects for a single
lookup, that broke the session report.

This fixes it, and also tweaks the schema so that the ID values that
come in the query string can be properly looked up as slugs or id's,
where previously it did not matter and now it does cause the queries
have been split.

Additionally, a now redundant function is removed, since the new one is
more generally useful (and otherwise does literally the same thing, but
in a more better-er way).