Include an image API to get the image type list
Since we have API's now that allow you to specify what kind of image
(if any) that you want in the output results, this includes an API
endpoint that provides a list of all such items and includes their
size, so that they can be presented in a UI.

The output does not contain any information as to the underlying variant
that actually sources the image, since that is a configuration detail
and could change. We also plan to have more than one image back end
eventually.