Fix issues inserting new games
After the schema cleanup, a whole was found in which we could not do a
game insert from BGG data because the new schema validation ensures that
the incoming metadata has slugs, but only when the user enters it.

The BGG routines were not including a slug, which caused things to fail.
This has been resolved now, as well as an accidental mis-comma accident
in the insert that also broke things.