Dump the dump
Since the code in the repo is more stable than during Devember, and
is actually being used in a production capacity, this shifts the dump
files and other data related to my specific game collection out to a
separate data folder, so that further changes don't need to be in the
repo.

The tool for dumping the database still exists, and has been moved into
a new tools folder.