Get ready for inclusion of authentication code
There has been some un-checked-in playing with potential authorization
flows, in the sense that we trust that the IAM system that we are using
(Cloudflare Access) will only allow authorized people through to the
application, but we want to know who they are.

To that end, we need to be able to validate the JWT that is provided and
make an API call as needed to look up user details for users that have
never been seen before.

This sets up the libraries needed, a simple tool for generating
development level tokens (which are not signed and thus not trusted
outside of localhost) and the configuration variables needed in order
to pull it off.

To that end, the contents of the wrangler file has been rewritten
somewhat to make it clearer what the various variables are used for,
etc.