Implement the start of our auth system
The largest part of the authorization system as a whole is entirely
handled by Cloudflare Access, which gates our users for us, blocks all
access, and provides to us in every request a JWT with their user UUID
of the actual user.

This code brings in the base needed to validate such token and then use
it to grab out of our database the record of who that particular user
happens to be.

This is smart enough to know when it is being accessed via a service
token, which gets mapped to a specific configured user. Currently that
is me but in the fullness of time that could instead be a root user of
some sort, though that is being held in abeyance until we get some
audit logs in place.

This call is currently not tied to anything, and it accesses the
database every time, which seems fast enough. The current idea is that
this could be invoked manually in order to get the user details for the
profile section of the UI, and internally invoked when we get to the
part of the development where user permissions are in place.

It should be fast enough to just always invoke for extra safety, though
at the current time I'm not sure that is strictly needed since the user
base is not open and Cloudflare Access prevents otherwise unauthorized
accesses from occurring.