Implement some user routines for search/insert
For our user authentication work that is upcoming, this includes some
database routines for looking up users by their external Cloudflare
Access user identifiers or their internal ID's, as well as inserting
a new user.

The insertion routine is meant to be invoked by automated code that can
determine the appropriate information for the record, and will be
invoked when an valid JWT token represents a user that is not currently
in the database.

The reasoning here is that the CF Access site is what gates who the
users that can connect are, so we only need to keep our internal
records up to date.

As a part of this, the uniqueness constraint on the firstName and
lastName combination on users was removed, because the main uniqueness
constraint should be the external ID.