Include an authorization endpoint in the API
This new endpoint only validates that we know who the user that is
accessing the API actually is.

The real work of allowing or disallowing access to the system is done
at the Cloudflare level, where users are specifically allowed or denied
access.

This endpoint uses the data in a JWT (which is fully validated) to
determine what user is accessing things, and will ensure that the
internal database of users is kept up to date by adding any new user
that does not yet exist.