Implement an initial user API
This starts the beginning of a simple user API which allows for fetching
the details of either a specific user, or getting the whole list of
users.

These are predicated on things the UI is going to need, and is still
pending on an API for adjusting some of the details of a user, such as
their name, email address, etc.