Include an initial household API
One of the features that has existed in the alpha since devember is a
table that contains information on the household(s) that a user is a
part of, which is a major part of being able to display information
such as wishlists and actually owned games collections (where currently
we just display all possible games, not just those that are owned).

This includes the start of an API to be able to visualize this
information.