Include households in the user information lookups
This is a small-ish change which:
  - Marks that a household can be the "primary" for a user
  - Includes a lookup to find the households a user is a part of
  - Includes that household list in the looked up user information

The information is added to the user object in two ways; first, the
primary household (if any) is added to a specific key, and then
secondly the whole list (including the primary) is included.

This allows for operations that want to know about the primary household
to not have to fetch that information on their own.