Update the code that authorizes the JWT
This is a small change that caches the lookup of users that are
accessing the application in an invariant in memory cache to save
execution cycles on subsequent accesses, since the external UUID never
changes.

This only caches the userId; no other user information is saved since
that is potentially variable.