Implement auth middleware and tie it to a user API
This change implements a new middleware that does JWT validations and
looks up the internal userId of the user making the request. The
middleware will store that userId into the context for downstream
handlers to use as they see fit.

To test this, and also for future UI purposes, there is a new /user API
for looking up details on the current user which uses this information
to do the lookup.