Alter the API function for fetching API data
This small change allows the client side API code to determine both its
own and the remote origin that it's been told about, so that it can
properly determine whether or not it should set the option that tells
the browser to ship credentials.

The default (for same origin) is to do so, but for different origins,
such as when we're working locally, this needs to be overridden.

Old man Odat doesn't want to subvert security mechanisms and cause more
CORS headaches than absolutely necessary.