Upgrade to the latest beta version of Zephyr
This updates the user interface code to use the latest version of Zephyr
(which is still in beta for a short period).

Largely this encompasses the notion that ws-x attributes are now
specified in a different manner and colors are expressed differently
than they were previously.

In addition, the main app HTML needed to not make the script a module
script, since that breaks Windstorm due to modules not having data
attributes accessible in the same way.

Lastly, there was an apparently long standing bug in the session list
code that was returning full size images and not thumbnails for games
in the list, which was being masked by the fact that the Windstorm size
rules were forcing the image smaller. Once those rules were broken, the
bug became more obvious.

With this commit, things should be largely styled as before, with the
exception of:
  - The buttons in the main menu now have a different outline color
  - The font sizes of various things seems to have changed