Implement an initial version of household details
This is a small set of changes that updates the routine that gets the
details of a game so that it can look up whether or not the game is
on the wishlist or owned (or both).

This data is not currently visualized in the details pane of the game.