Rework our internal API fetch functionality
This modifies our internal code that talks to the API so that when pages
and components need data, they can make a simple `api.get()` call or
similar, and get back the data result directly.

The new api fetch call handles all errors, including when the result
that comes back is incorrectly formatted. As a result, it returns the
'data' field directly back, which cleans up calling code somewhat.

We also include in this change actual error handling such that if any of
the api calls fails, the LoadZone will show an error message.