Use the new User store for the main app
This is a small change, but we wrap the main router in the page with a
LoadZone that ensures that the user data has been fetched before it
allows the page content to appear.

This also disables the button that will eventually become the menu, with
the proviso that it cannot also do that for the link button, but the
link button will become a profile button that allows for taking
actions, and we want that to remain enabled even if the load fails so
that the logout link can be clicked there instead.