Display ownership and wishlist status in details
This is a small visual change, but when displaying the core details of a
game, there is now a distinguisher for whether or not this game is
owned or on the game wishlist.

Games that are owned have a star icon, games that are wishlisted have a
heart, and games that are neither have a slashed through star.

In addition to that, if the game is owned or wishlisted the main name
displayed is the one that is owned/wishlisted; as such we now also
dispay the entire AKA name list, since we're lazy and it's now possible
that the first name is not the one that is visualized.