Slightly reconfigure the slide out menu content
In order to make the application more navigable on phones and other
small screens, the links in the two menus are now set up to style
themselves as buttons instead.

There was also a bit of misc renaming of the text, and all of the places
that display the user name also display the user household.