Set the GameName field to be case insensitive
At great personal cost to my time and sanity, this change adjusts the
schema so that the textual names that are assigned to games are not
case sensitive for the purposes of insertions and comparisons.

This is needed for some upcoming API work in which we want to
dynamically adjust the list without using lots of table scans to try
and find existing entries by their text.