Include a new API to look up game names for a game
Upcoming UI work requires a method to get at all of the names that a
game is known by, including their ID values.

This data is not available from the details page since it's not needed
there, and adding it would force other UI elements that want to know the
name list to do heavyweight game lookups.

Like our other API's, this one allows you to look up the game by either
it's ID value or it's slug.