Refactor the owned games endpoints for better name
Since the button for showing the list of owned games calls it the game
collection, and we're about to add in more endpoints that deal with the
collection, this splits that request out into a subset so that the
layout of the folder matches what we do elsewhere.

As a part of this, the route changed, so the UI was updated as well.