Resolve an issue with game expansions
Resolve a problem where inserting expansion records for games that are
not in the database causes the expansion insert to fail.

This seems like it was caused by some of the zod refactoring, because D1
was getting janked about a value being undefined and not null, which
seems suspiciously like something that would have been validated
previously but now is now.