Include API for adding/removing from a wishlist
This commit includes the database routines, validation schemas and
request methods to allow for adding or removing a game from a
household's wishlist.

This also tweaks the previous commit so that when you add a game to the
collection, if it's in the wishlist, it is automatically removed, since
places that don't do that are extremely irritating.