Fix fetch API wrapper for non-get request w/a body
Due to a terrible miscalculation related to not having any such requests
in the API at the time, the client side wrapper for talking to the API
did not correctly handle non-get requests with a body.

In particular, it did not encode the body object as a JSON string in
that case, and did not set an appropriate content type.