Implement UI controls for collections/wishlists
This commit implements the appropriate UI structure for allowing the
previously created buttons for adding and removing a game from the
collection and the wishlist so that they are actually implemented.

Additionally this slightly changes how the page is generated so that as
the state of the collection changes, we can cause the page to reactively
change.