Make game details include full name details
Previously, the names field in a game details object was just the
textual names for the game and nothing else, with the first item in the
list being the "primary" name that the game is known by (for which
there is a guarantee that there is always exactly one).

This alters the call to the DB so that the returned values include not
just the name but also the ID and the indication as to whether or not
the name is primary.

In addition, since the primary name is so commonly used throughout the
UI, the returned value includes the textual version of the name as a
primaryName field for easier access.

This required some small changes in the UI where it used to reference
the name list as text; we also double checked and this does not seem to
affect anything in the back end.