Update the owner and wishlist data endpoints
This is a small change that updates the data that we capture for the
owner and wishlist data such that it is consistent with the data that
is used for similar purposes elsewhere in the object.

For example, the owned publisher is an actual publisher object and not
just some additional fields.

This required some small tweaks to the UI to get done.