Refactor the BGG "database" code
This swaps the BGG code out of the db dir and into the lib dir instead
since it does not actually do anything with the database; the signature
of the main exported function was also modified to be more readable.

In addition, since a recent change to the internal game details endpoint
included us inlining the primary name for clarity in the structure,
this now also does the same so that the two are a close correlation.