Refactor the image code
Rename all of the internal DB methods that get and manipulate the
image data to have nicer names.

As a part of this split, the code in the db/image.js file that is not
actually related to working with the database in general has been bumped
out into a lib/image.js file instead.