Split the userhousehold code out
Small change here; shift the method that finds all of the users in a
household into the household code, and the code that finds all of the
households of a user into the user code.

Technically both could be in either file since they straddle both sides
of that data plane, but this makes the intent a bit clearer.