Switch build style to get around CF caching issues
This swaps the build of the application such that instead of compiling
the JS directly to an app.js file and using a static HTML file, we now
use a plugin for rollup written by Axel that sources the HTML file and
uses a specific tag within to know where the JS source is.

This allows the build to generate an app.js that has an asuid as a part
of the name.

This doesn't change anything in the app, but does help to resolve an
issue whereby CloudFlare aggressively caches JS files (but not HTML
files).

Up until now, the cache period was short enough that a simple local
flush was enough to resolve the issue, but that seems to no longer be
the case.