Alter the import location of the client API
While working on some potential API changes, it seems like it would be a
good idea to abstract away the way that the API import works so that we
can spread the API over several files rather than just a single file.

As such, this abstracts the API import so that it is not specific to any
one file, allowing us to easily alter the underlying mechanism without
any outward facing changes.