Implement a simple add dialog for collect/wishlist
This is a bit of a dirty proof of concept that could use a fair bit more
refactoring, but it works and can be cleaned up later.

We have a simple dialog that prompts to the name and publisher when
adding a game to the owned list; the same dialog is used when adding
games to a wishlist as well, although there only a name is used.