Refactor the client API into a cohesive object
The original API code was a quick PoC, which got extended in a recent
commit as we worked towards cleaning up the API structure a little bit
to make imports nicer.

This finalizes that job that making a single API object that contains
not only the raw API for use as needed but also various namespaced
wrapper methods that enhance readability and hide the underlying
details of what is happening.