Fix issue with image upload env duplication
An issue existed in which, since the single CloudFlare images account is
used for both production and development, a collision could happen when
adding in a game in both environments because the image upload fails
when told to upload a file that already exists.

This change introduces new functionality that allows for looking up the
image that would otherwise be uploaded, and if found instead of trying
to upload, the details from the existing upload are used instead, as if
an upload just happened.