Augment our builds to be production ready
This updates our build infrastructure for the client side code such that
when a build happens it can determine if it is a production or
development build and act accordingly.

For the time being, the difference between the two is related to using
the terser() plugin in a production build, so that the release versions
have smaller output that downloads faster.