Set up to show commit hash in UI releases
Include a new tool script that can gather the commit hash of the most
recent commit and use it to generate a JS file that allows for code to
load that commit data.

This is triggered by the UI code build scripts and injected into the
built files so that it can be displayed in the UI.

Since the UI code deployment runs on the CloudFlare build server in a
chroot jail on the client folder after being checked out, we include a
symlink to the tool script so that it can be executed during the build.

I am not 100% sure this will actually work, so more mojo may be
required...

In any case, a new component allows for linking to the git commit in
question.

If this works, we will augment the API code to also have a version API
that can retrieve similar data from the back end.