Resolve issue with expansion
At some point in the past after the Zod refactoring, there was an issue
exposed where if a base game id was undefined, the insert failed. This
worked prior to zod because we had different custom internal
validation.

This fixes the same error, but on the expansion side of the operation
instead.