Upgrade the used Zephyr version
Upgrade to the latest release version of Zephyr, which also brings with
it a new version of Windstorm, which also brings with it bootstrap
icons in favor of Tabler icons since they make the build smaller.

As such, this ALSO includes icon tweaks so that things still display as
expected.