Update the release tracker to generate an object
For upcoming changes, it makes more sense for our script that grabs the
current commit so that it generates an object that includes not only the
commit hash but also the date on which the file was generated and an
indication as to whether or not the tree was clean at the time.

This also includes a change to the server side code so that when
wrangler bundles or runs the dev server, it also runs our script.