Use tabs for the main content
In order to tighten up the display a bit, this change modifies the main
part of the game details page such that there are tabs to swap between
images, files and the description of the game.

Largely only the image is interesting in this context (and files are not
yet supported), so this makes the main page more compact.