Streamline the secondary content
Continuing on with the previous commit, this alters the larger bulk of
the remaining data in the details pane such that it is displayed in
tabs like the main data is.

A new component lists expansions as a table rather than an unwieldy
list, which can be augmented in the future to include the session list
inline rather than it being a distinct page instead.

Additionally, the buttons for adding and removing to the collection have
been left alone since they are currently not guarded against accidental
presses.