Rename the existing Wishlist Table
We're working towards having more than one wishlist available per
household instead of the single one that the code currently expects.

As a first phase of this:
  1. Rename the existing Wishlist table to WishlistContents
  2. WishlistContents now has a wishlistId field to track the list entry
  3. WishList becomes a new table that is the list of all wishlists

The code is also modified to pull wishlist data from the new content
table and to store/delete it from there, currently using a hard coded
wishlistId to make the transitional code easier to handle.