Clean up API endpoints for current wish list API
The wish list API is currently in a state of evolution as we add the
power for there to be more than one wish list per household rather than
just a single one.

This alters the existing wish list endpoints to have names that make
more sense for adding and removing items in a wish list, as well as
getting the overall contents of a wish list.

The endpoint names are changed, but their parameters remain as they
currently are; in particular the items always work with the wishlist
whose ID is 1.