Shift the list of game play types to a table
A semi common task (in that it has now happened twice since the
inception of Luddatumbazo) is that of augmenting the list of possible
places where games can be played, as far as the session reporting tools
go.

This is a giant pain in the ass because the play type is a CHECK
constraint, which makes altering the schema of the file problematic
since it requires dropping and recreating all of the session data.

Instead of that, this now includes that data in a separate table in the
database, along with a new API endpoint that can request it.