Update the client side to use new gameType API
This is a small-ish change; the list of game types is static at runtime,
so the client side only needs to request once at initial page load.

This has been loaded into the call in the server store, since this is a
form of server data. The store is modified to make a second call and to
have some extra methods for doing lookups, and the client code is now
smaller in both places (and more robust) by using the methods, which
themselves are using dynamically sourced data.