Update the UI to allow for multi wishlist display
This alters the back end API to require that the call to get the
contents of a wish list contain not only an owning household ID but also
the wish list identifier.

The request was updated to also require this, but to allow for the
request to be either a numeric ID or a slug.

With these in place the client side API wrapper was modified to require
that a wishlist id or slug is called, and the main wishlist page now
uses the result of a wishlist list query to first get the list of known
wishlists and then also uses it to fetch just the required wishlisted
games.