Include an API for inserting a new household
Previously the one and only household was hand created at the time that
the database was populated.

The recent changes to wish lists are predicated on the idea that every
household will always have at least one wish list, which is flagged as
being the root playlist.

In order that 6 months from now when we build out the UI for adding a
household we don't screw stuff up by forgetting we need to also create
a wish list, this includes a new API for doing exactly that.