Details now display the wishlist a game is in
The back end queries for pulling information on the ownership and wish
list status of a game were updated so that the return values have more
sub keys. In addition (and primarily) the wish list data record in the
game details now also includes the wish list that the game is on.

The details pane was adjusted to show this information directly.