Update the household ID schema parameter
While working on the wish list code we needed to create a new schema for
dealing with actual wish lists, and that required a wish list id or
slug type parameter.

This required us to augment that schema by renaming the idOrSlug value
since there were effectively two of them.

This completes that action for the collection and other household
related items, since they were still using the old name and thus looked
out of place.