Update household handling in the UI
There were various places in the user interface where we were not
properly handling cases where a user might not have a household.

This has now been remedied in all known locations; items that require a
household now have their links blocked so that they can't be clicked,
as well as having guards to make sure that manual page URL's also work.

The API functions that require a household were also adjusted to take an
actual household as an argument rather than a user, paving the way for
users that belong to more than one household.