Fix the terrible multi loading wishlist hack code
Fix the code where we were hackily loading the list of wish lists for a
user for every game details page load so that instead the list of lists
is loaded once when the page first loads and then the value is
persisted thereafter.

This is done via a store.