Include a DB routine to record an uploaded asset
Include in the database a new table for tracking files that have been
uploaded to the R2 bucket for the purposes of attaching files to
specific games, and include a new DB routine that can UPSERT records
for uploads into the table.

We treat the bucket key as the unique factor, and have an ON CONFLICT
resolution clause on any insert that we do that allows us to just
update any existing record rather than failing if we violate the
uniqueness constraint.