Include the ability to delete an asset
Allow for the deletion of an asset, both from the DB and from the R2
bucket, based solely on the key in question.

This allows the delete to happen even if the key isn't in the bucket or
isn't in the database, allowing for cleanup even if things get desynced
somehow.

The endpoint considers it an error to attempt to delete a key that does
not exist in either location.