Clean up the package files for the project
This cleans up the package.json files throughout the repo to make the
scripts nicer, and also finally implements a wrangler.toml file for the
client part of the site.

Back late last year when we were investigating using CF Pages, the
support for a wrangler config file for a pages dev site was not there
yet, but this seems to have been fixed now, though the documentation
does still say that it's in beta.