Update the asset schema to include description
Something that was left out of the original schema for file assets was a
descriptive string of text that describes what the file actually is/is
for, etc.

This is a small change to the schema and related API to ensure that the
description passes through the API; it is considered non-optional,
although nothing prevents you from using a useless description if you
see fit.