Include file lists in the game details pane
Include the initial implementation of the ability of the client side
code to display a list of all of the files that have been attached to a
particular game.

This is not 100% complete yet, since it fetches the file list every time
the tab is focused. However, we also load the game details and game
list on every page load too, so meh.