Update the SessionList to control game display
For an upcoming change we would like to inline the session report table
directly into the game details pane. In doing so the table would not
need to link to the game since it's already on that page.

In addition we require the ability to style the color of the component
and the number of rows displayed based on where it's used.

Currently, as an extra awful hack, if games are not being displayed the
page size will try to go to a minimum of the number of loaded rows and
the page size provided, so that when it is used in game details pages
it can constrain its size.

This last part should actually be being done in a better way, but we're
still feeling out the overall page design.