Update how the file asset table works
This is a small WIP to change how the display of files in the asset
table is visualized.

Instead of showing the date, we instead show an icon that pops up a
dialog that has file information within it.

The dialog is still a bit crude, but for the purposes of testing, it is
enough to show the data that we want the dialog to populate.