File dialog cuts off close button
The file information dialog had an issue where the size of the dialog
content did not grow wide enough to convey the whole content of the
dialog, which cuts off the close button on smaller mobile devices and
makes closing the dialog impossible. This is now fixed.

In addition, the dialog has been updated so that it will close whenever
you click anywhere outside of it.