Implement an IconButton
This implements an IconButton custom component that wraps a Zephyr Icon
and Button combination, and allows for the caption and/or icon to
change for a configurable period of time when the button is pressed, if
desired.

The main use case for this is in the FileInfoDialog, where we have a
button that copies a link to the current file to the clipboard and we
would like to have a bit more visual "oomph".

This is also used in several other places, to keep all of the buttons
that have icons similarly styled.

It would be handy to have a secondary version of this (or a more
configurable one) for IconEntryButton, since that is used for some
things in the UI as well. For now, those are not handled this way.