Update to using Zephyr version 0.4.5
Update the used version of Zephyr to the latest release, which is 0.4.5
and brings with it a few new style options and many new components.

For our purposes here the chief differences are in how the overall
theme, tabs and selects are configured.

In addition there is also a new AsyncButton component that makes life
better in regards to disabling a button automatically when an async
call is happening, a new confirmation dialog that we use to make
sure that the user is OK with removing things from their wishlist
and/or collection, and a toast mechanism to provide extra feedback when
actions are taken.