Prepare to use Zephr as the application router
We are preparing to modify all of the routing in the app so that we use
the router built into Zephyr rather than svelte-spa-router for the job.

This small commit removes all direct uses of that library except the
main one in the App.svelte file that is responsible for all of the
routing.