Utilize new Zephyr functionality for confirmations
One of the new features in Zephyr is a confirmation dialog box that
makes it easy to show a message asking the user to confirm an action
with a yes or not response, and then act accordingly.

As a test, this is now added to the functionality that performs a
removal from a wishlist or the overall game collection, which
previously allowed you to destroy data if you fat fingered the button.