Coalesce the wishlist specific DB code out
Previously the wishlist specific DB code existed within the household DB
file because wishlists are related to households.

However, once we fleshed out the API for wishlists to allow for
multiples, and the addition/removal of them, the code became more
unclear on separation of concerns.

This change fixes that by pulling the wishlist specific functions out
into their own file.