Small code cleanups
Include some small code cleanups that were originally implemented to be
a part of some backup code that eventually never got committed in this
repository because we wanted it to be a separate project instead.

This includes things like fail() being able to return a body, ensuring
that DB results will display correctly even if there is no time,
changing the name of the bound databases and buckets, and ensuring that
all of our local dev versions persist data for D1 and R2 into the same
location so they can all share local data.