Change how files keys for R2 are generated
Something learned during work on Sekurkopio is that rclone does not like
R2 keys to start with a leading `/` since it strips that off of paths
when you tell it where to look, then gets angry that what it came up
with does not exist in the bucket.

For more clarity, this changes the keys used for file uploads and,
consequently, what is used for generating links.