Minor code cleanup
Recent changes to how we use Wrangler locally allow us to have all of
our wrangler local dev versions running with a shared state folder so
that when one of them writes to e.g. a database, others can see it.

Doing this requires all wrangler commands to be told where to persist
data to. So, for the time being, redact the ability to seed a local DB
from the project, since that is not going to work out well.

In addition to that, a couple of minute cleanup tweaks.