Reorganize the project
Since this was originally developed there have been some enhancements in
the realms of Cloudflare Workers and Pages, and also I now think that
the file structure is lame.

So, we are reorganizing everything to have a centralized package.json
file and to use Cloudflare WOrkers Assets instead of pages so that the
application bundles its own content all in one build, which makes
deployments and some configurations less irritating.