Clean up the build to remove currentCommit.js
Previously we had a script that would execute to grab for us the current
commit hash, the current date and whether or not the source tree was
dirty or not, and create a file out of it so that it could be used in
the server and client.

We don't want to do that since we now have the client and server being
part of the same overall deployment, so it makes somewhat less sense
now.

Instead, the rollup config for the client now takes care of this on its
own since it does not need to be in both places. This also means that
the server endpoint for this has been redacted as well.