Swap all of the API's over to the new format
This rather hefty change that will make us sad in the future no doubt
because it's so large converts all of the original Hono route
definitions with those for the hono file route plugin.

Functionality here should be identical to what was present previously,
with the exception that the common routines are coming from the new
@odatnurd/cf-requests package rather than the internal common code.

The only thing not migrated over is the validation, which still uses
Zod schemas for now.

As such, everything has been more clearly marked as using the legacy
Zod validation routines to make it clearer; we can move them over
piecemeal as needed.